chore: add success message & rename method

+++ src/main/java/com/example/demo/dto/RequestScheduleDTO.java
@@ -8,7 +8,7 @@ import java.time.LocalDateTime;
 
 @Data
 public class RequestScheduleDTO {
-    private int id; // 스케줄 아이디
+    private Long id; // 스케줄 아이디
     private String title; //스케줄 제목
     private String content; // 스케줄 내용
     private String color; // 스케줄 색상
